+++ src/App/AppHeader/index.js
@@ -6,6 +6,7 @@ import { GithubLogo, SlackLogo } from "./HeaderIcons";
 import pkg from "~/package.json";
 
 const basename = process.env.basename;
+const brand = process.env.brand;
 
 const AppHeader = () => (
     <header className="topbar topbar-md-wide designguide-header">
@@ -13,7 +14,7 @@ const AppHeader = () => (
             <i className="material-icons topbar-btn-icon">menu</i>
         </button>
         <a href="/" className="topbar-logo">{"\n"}
-            <img src={`${basename}img/spay_horizontal_pos.svg`} alt="logo" />{"\n"}
+            <img src={`${basename}img/${brand}-logo.svg`} alt="logo" />{"\n"}
         </a>
         <nav className="topbar-nav">
             <div className="topbar-link-container">
@@ -25,7 +26,7 @@ const AppHeader = () => (
                 <NavLink to="/examples" activeClassName="active"><span>Examples</span></NavLink>
                 <div className="topbar-info topbar-link-right">
                     <div className="topbar-info-contact">
-                        <a href="https://github.com/SwedbankPay/design.swedbankpay.com" target="_blank" rel="noopener noreferrer">
+                        <a href={`https://github.com/${brand}/design.${brand}.com`} target="_blank" rel="noopener noreferrer">
                             <GithubLogo />
                         </a>
                         <a href="https://payex.slack.com/messages/C0L3W8B2S/" target="_blank" rel="noopener noreferrer">
@@ -33,7 +34,7 @@ const AppHeader = () => (
                         </a>
                     </div>
                     <div className="topbar-info-version">
-                        <a href={`https://github.com/SwedbankPay/design.swedbankpay.com/releases/tag/${pkg.version}`} target="_blank" rel="noopener noreferrer">v.{process.env.version || pkg.version}</a>
+                        <a href={`https://github.com/${brand}/design.${brand}.com/releases/tag/${pkg.version}`} target="_blank" rel="noopener noreferrer">v.{process.env.version || pkg.version}</a>
                     </div>
                 </div>
             </div>