+++ src/payex.js
@@ -1,0 +1,18 @@
+import React from "react";
+import { render } from "react-dom";
+import App from "./App";
+
+import "prismjs";
+import "prismjs/themes/prism.css";
+
+import "./less/payex.less";
+import "./less/documentation.less";
+
+render(
+    <App />,
+    document.getElementById("designguide")
+);
+
+if (module.hot) {
+    module.hot.accept();
+}