+++ src/scripts/main/utils/index.test.js
@@ -1,0 +1,68 @@
+import { extendObj, isWithinBoundingBox, handleScrollbar } from "./index";
+
+describe("scripts: utils", () => {
+    describe("- extendObj", () => {
+        it("extends current object with new object", () => {
+            const oldObj = { test: "test" };
+            const newObj = { test2: "test2" };
+            const expectedObj = {
+                test: "test",
+                test2: "test2"
+            };
+
+            const extObj = extendObj(false, oldObj, newObj);
+
+            expect(expectedObj).toEqual(extObj);
+        });
+
+        it("does not overwrite property when overWrite is set to false", () => {
+            const oldObj = { test: "test" };
+            const newObj = {
+                test: "test1",
+                test2: "test2"
+            };
+            const expectedObj = {
+                test: "test",
+                test2: "test2"
+            };
+
+            const extObj = extendObj(false, oldObj, newObj);
+
+            expect(expectedObj).toEqual(extObj);
+        });
+
+        it("overwrites property when overWrite is set to true", () => {
+            const oldObj = { test: "test" };
+            const newObj = {
+                test: "test1",
+                test2: "test2"
+            };
+            const expectedObj = {
+                test: "test1",
+                test2: "test2"
+            };
+
+            const extObj = extendObj(true, oldObj, newObj);
+
+            expect(expectedObj).toEqual(extObj);
+        });
+    });
+
+    test.todo("Write tests for isWithinBoundingBox");
+
+    describe("- handleScrollbar", () => {
+        it("adds class has-vscroll to body if innerwidth is larger than clientwidth", () => {
+            handleScrollbar();
+
+            expect(document.body.classList).toContain("has-vscroll");
+        });
+
+        it("removes has-vscroll if vertical scrollbar does not exist", () => {
+            document.body.classList.add("has-vscroll");
+            window.innerWidth = 0;
+            handleScrollbar();
+
+            expect(document.body.classList).not.toContain("has-vscroll");
+        });
+    });
+});